apps/autopulse-agent/requirements.txt
requests
pyyaml
requests==2.31.0